fix: Move options around
